% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\docType{class}
\name{CovarFun-class}
\alias{CovarFun}
\alias{CovarFun-class}
\title{Class CovarFun.}
\description{
Class \code{CovarFun} defines a covariance function to be used as part of
a \code{GPC} gaussian process classifier. A \code{CovarFun} object is made
up of a kernel k(x,y), a set of hyperparameters used in the kernel, and
a function returning the gradient of the kernel with respect to each of the
hyperparameters.
}

