% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\name{covarFun.LatentPlusNoise}
\alias{covarFun.LatentPlusNoise}
\title{Augment \code{CovarFun} with latent function and noise hyperparameters}
\usage{
covarFun.LatentPlusNoise(covarFun)
}
\arguments{
\item{covarFun}{\code{CovarFun} object to augment.}
}
\value{
covarFun augmented with lsf, lsn hyperparameterss for signal and
 noise magnitude.
}
\description{
Creates a covariance function k which augments some covariance function k_f
with two extra hyperparameters lsf and lsn such that:
k(x, y) = exp(lsf)*k_f(x, y) + exp(lsn)*I[x=y]
Automatically provides derivatives of lsf, lsn. Used for neater covariance
funtion specifications. lsf, lsn stand for log(sigma^2_f) and log(sigma^2_n),
where 'f' labels the magnitude parameter and kernel for the latent
function, whilst 'n' labels the magnitude parameter for the noise.
}
\examples{
C = covarFun.LatendPlusNoise(covarFun.SE(0))
}

