% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\name{covarFun}
\alias{covarFun}
\title{Construtor method of CovarFun class}
\usage{
covarFun(k, dk, hp)
}
\arguments{
\item{k}{A kernel function (object, x, y) -> numeric() which, given data x
and y returns an inner product. Kernel hyperparameters may be accessed with
 object@hp.}

\item{dk}{A function (object, x, y) -> list() which returns the gradient of
 k with respect to the hyprparameters in the form of a lsit of the same
 shape as the kyperparameters}

\item{hp}{A list of kernel hyperparameters.}
}
\description{
Creates a new CovarFun object intended to be used inside a \code{GPC}
Gaussian Process Classifier.
}
\details{
A \code{CovarFun} object extends the kernel object which supplies a kernel
function along with a list of hyperparameters. \code{CovarFun} also supplies
a function returning the gradient of the kernel with respect to the
hyperparameters, such that the hyperparameters may be tuned by the \code{GPC}
class.
}
\examples{
# Isotropic squared exponential covariance function with log length scale
# (ll) hyperparameter
library("gpclassifier")
k  = function(.Object, x, y) {
  exp(-0.5 * sum(exp(-2*.Object@hp$ll) * (x-y)^2))
}
hp = list(ll=0)
dk = function(.Object, x, y) {
  list(ll=.Object@k(.Object, x, y) * exp(-2*.Object@hp$ll)*crossprod(x-y))
}
C  = covarFun(k, dk, hp)
}

