% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\name{covarFun.SE}
\alias{covarFun.SE}
\title{Squared Exponential covariance function.}
\usage{
covarFun.SE(ll = 0)
}
\arguments{
\item{ll}{log length scale hyperparameter.  \code{ll} must be of length 1 or
the same length as the input vectors. If length 1 then this is an
isotropic SE covar fun, else there is a length scale for each dimension
of the input.}
}
\value{
\code{CovarFun} for a squared exponential covariance function
}
\description{
Squared Exponential covariance function.
}
\examples{
C <- covarFun.se(0)
}

