% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/siteParams.R
\name{EP}
\alias{EP}
\title{Calculate site parameters, likelihood and likelihood gradient using
Expectation Propagation}
\usage{
EP(object)
}
\description{
Calculate site parameters, likelihood and likelihood gradient using
Expectation Propagation
}

