% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPC.R
\name{getCovarFun}
\alias{getCovarFun}
\title{GPC Return the covariance function (class \code{CovarFun}) used.}
\usage{
getCovarFun(object)
}
\description{
GPC Return the covariance function (class \code{CovarFun}) used.
}

