% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPC.R
\name{getDLml}
\alias{getDLml}
\title{Return partial derivatives of the log marginal likelihood
with respect to each of the hyperparameters.}
\usage{
getDLml(object)
}
\description{
Return partial derivatives of the log marginal likelihood
with respect to each of the hyperparameters.
}

