% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPC.R
\name{getK}
\alias{getK}
\title{GPC Return the covariance matrix; the matrix of inner products
between each pair of data points in the space induced by the covariance
function.}
\usage{
getK(object)
}
\description{
GPC Return the covariance matrix; the matrix of inner products
between each pair of data points in the space induced by the covariance
function.
}

