% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\name{getKernel}
\alias{getKernel}
\title{Return kernel function k:x,y -> numeric(). Differs from the kernel function
specified when constructing the \code{CovarFun}, since the kernel function
returned only requires parameters x,y, not object.}
\usage{
getKernel(object)
}
\description{
Return kernel function k:x,y -> numeric(). Differs from the kernel function
specified when constructing the \code{CovarFun}, since the kernel function
returned only requires parameters x,y, not object.
}

