% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\name{getKernelGrad}
\alias{getKernelGrad}
\title{Return kernel gradient function k:x,y -> list(). As with
getKernel(CovarFun), this differs from the kernel gradient function
specified when constructing the \code{CovarFun}, since the function
returned only requires parameters x,y, not object.}
\usage{
getKernelGrad(object)
}
\description{
Return kernel gradient function k:x,y -> list(). As with
getKernel(CovarFun), this differs from the kernel gradient function
specified when constructing the \code{CovarFun}, since the function
returned only requires parameters x,y, not object.
}

