% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPC.R
\name{GPC}
\alias{GPC}
\title{Class GPC.}
\usage{
GPC(X, Y, covarFun = NA)

GPC(X, Y, covarFun = NA)
}
\arguments{
\item{X}{Matrix of input data; sample in rows.}

\item{Y}{Logical vector of binary labels.}

\item{covarFun}{Covariance function to use. Must be of class CovarFun. If
 omitted, the squared exponential covariance function is used by default.}
}
\value{
S4 object of class GPC, where covarance function hyperparameters
 have been set to their maximum likelihood estimates.
}
\description{
Class \code{GPC} defines a Gaussian Process Classifier. GPC(),
GPC(X, Y, covarFun) creates a new GPC object used to predict labels for
new input data.

Construct a new GPC object.
}
\details{
Covariance function hyperparameters are selected automatically through
maximum likelihood. This class is implemented using the Expectation
Propagation approximation detailed in (Gaussian Processes for Machine
Learning, Rasmussen and Williams, 2006).
}
\examples{
# Create synthetic dataset
X <- matrix(rnorm(60), ncol=2)
Y <- rowSums(X^2) < 1

# New GPX Object with default squared exponential covariance function.
gpc <- GPC(X, Y)

# Predict labels for new data
Xst <- matrix(rnorm(60), ncol=2)
Yst <- predict(gpc, Xst)
}

