% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hypParams.R
\name{hpTune}
\alias{hpTune}
\title{Return maximum likelihood covariance function hyperparameters}
\usage{
hpTune(object)
}
\description{
Return maximum likelihood covariance function hyperparameters
}

