% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPC.R
\docType{methods}
\name{predict,GPC-method}
\alias{predict,GPC-method}
\title{Predict labels from unlabelled data.}
\usage{
\S4method{predict}{GPC}(object, Xst)
}
\arguments{
\item{object}{A fitted GPC objet}

\item{Xst}{A matrix of data for which to produce label predictions}
}
\description{
Use a trained GPC classifier to produce label predictions of a new set of
unlabelled data.
}
\examples{
# Create synthetic dataset
X <- matrix(rnorm(60), ncol=2)
Y <- rowSums(X^2) < 1

# New GPX Object with default squared exponential covariance function.
gpc <- GPC(X, Y)

# Predict labels for new data
Xst <- matrix(rnorm(60), ncol=2)
Yst <- predict(gpc, Xst)
}

