% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CovarianceFunction.R
\name{setHP<-}
\alias{setHP<-}
\title{Change the \code{CovarFun} hyperparameter list to the list supplied.}
\usage{
setHP(object) <- value
}
\description{
Change the \code{CovarFun} hyperparameter list to the list supplied.
}

