% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GPC.R
\name{update<-}
\alias{update<-}
\title{Update \code{X}, \code{Y} or \code{covarFun}}
\usage{
update(object) <- value
}
\description{
Update the data \code{X}, labels \code{Y} or covariance function
\code{covarFun}, causing a recalculation of hyperparameters,
site parameters and covariance matrix.
}

